Removed duplicate MapLocations code & added missing license headers
