--- runelite-client/src/main/java/net/runelite/client/plugins/spelleffecttimers/MapLocations.java
@@ -1,2113 +1,0 @@
-package net.runelite.client.plugins.spelleffecttimers;
-
-import net.runelite.api.Constants;
-
-import java.awt.*;
-import java.awt.geom.Area;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-public class MapLocations
-{
-	private static final List<Shape>[] DEADMAN_SAFE_ZONES = new List[Constants.MAX_Z];
-	private static final List<Shape>[] PVP_WORLD_SAFE_ZONES = new List[Constants.MAX_Z];
-	private static final List<Shape>[] WILDERNESS = new List[Constants.MAX_Z];
-
-	private static Area getArea(List<Shape> shapes)
-	{
-		Area area = new Area();
-		for (Shape shape : shapes)
-		{
-			area.add(new Area(shape));
-		}
-		return area;
-	}
-
-	private static Area getArea(List<Shape> shapes, Rectangle view)
-	{
-		Area area = new Area();
-		for (Shape shape : shapes)
-		{
-			if (shape.intersects(view))
-			{
-				area.add(new Area(shape));
-			}
-		}
-		return area;
-	}
-
-
-	public static Area getDeadmanSafeZones(int plane)
-	{
-		return getArea(DEADMAN_SAFE_ZONES[plane]);
-	}
-
-	public static Area getPvpSafeZones(int plane)
-	{
-		return getArea(PVP_WORLD_SAFE_ZONES[plane]);
-	}
-
-	public static Area getWilderness(int plane)
-	{
-		return getArea(WILDERNESS[plane]);
-	}
-
-	private static void initializeWithEmptyLists(List<Shape>[] array)
-	{
-		Arrays.fill(array, new ArrayList<>());
-	}
-
-	static
-	{
-		initializeWithEmptyLists(DEADMAN_SAFE_ZONES);
-		initializeWithEmptyLists(PVP_WORLD_SAFE_ZONES);
-		initializeWithEmptyLists(WILDERNESS);
-
-		defineDeadmanSafeZones();
-		definePvpSafeZones();
-		defineWilderness();
-	}
-
-	private static void defineDeadmanSafeZones()
-	{
-		// Varrock
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				3182, 3382,
-				3182, 3399,
-				3174, 3399,
-				3174, 3448,
-				3198, 3448,
-				3198, 3449,
-				3197, 3449,
-				3197, 3450,
-				3196, 3450,
-				3196, 3451,
-				3195, 3451,
-				3195, 3452,
-				3194, 3452,
-				3194, 3453,
-				3193, 3453,
-				3193, 3454,
-				3192, 3454,
-				3192, 3455,
-				3191, 3455,
-				3191, 3456,
-				3190, 3456,
-				3190, 3457,
-				3185, 3457,
-				3185, 3463,
-				3186, 3463,
-				3186, 3464,
-				3187, 3464,
-				3187, 3467,
-				3167, 3467,
-				3167, 3468,
-				3163, 3468,
-				3163, 3467,
-				3142, 3467,
-				3142, 3468,
-				3141, 3468,
-				3141, 3469,
-				3140, 3469,
-				3140, 3470,
-				3139, 3470,
-				3139, 3471,
-				3138, 3471,
-				3138, 3484,
-				3139, 3484,
-				3139, 3485,
-				3140, 3485,
-				3140, 3486,
-				3141, 3486,
-				3141, 3491,
-				3140, 3491,
-				3140, 3492,
-				3139, 3492,
-				3139, 3493,
-				3138, 3493,
-				3138, 3515,
-				3139, 3515,
-				3139, 3516,
-				3140, 3516,
-				3140, 3517,
-				3141, 3517,
-				3141, 3518,
-				3160, 3518,
-				3160, 3517,
-				3161, 3517,
-				3161, 3516,
-				3162, 3516,
-				3162, 3515,
-				3167, 3515,
-				3167, 3516,
-				3168, 3516,
-				3168, 3517,
-				3169, 3517,
-				3169, 3518,
-				3191, 3518,
-				3191, 3517,
-				3192, 3517,
-				3192, 3516,
-				3193, 3516,
-				3193, 3515,
-				3194, 3515,
-				3194, 3514,
-				3195, 3514,
-				3195, 3513,
-				3196, 3513,
-				3196, 3512,
-				3197, 3512,
-				3197, 3511,
-				3198, 3511,
-				3198, 3510,
-				3199, 3510,
-				3199, 3509,
-				3200, 3509,
-				3200, 3508,
-				3230, 3508,
-				3230, 3507,
-				3231, 3507,
-				3231, 3506,
-				3232, 3506,
-				3232, 3505,
-				3233, 3505,
-				3233, 3504,
-				3234, 3504,
-				3234, 3503,
-				3235, 3503,
-				3235, 3502,
-				3252, 3502,
-				3252, 3496,
-				3253, 3496,
-				3253, 3495,
-				3254, 3495,
-				3254, 3494,
-				3255, 3494,
-				3255, 3493,
-				3263, 3493,
-				3263, 3472,
-				3264, 3472,
-				3264, 3471,
-				3265, 3471,
-				3265, 3470,
-				3266, 3470,
-				3266, 3469,
-				3267, 3469,
-				3267, 3468,
-				3268, 3468,
-				3268, 3467,
-				3269, 3467,
-				3269, 3466,
-				3270, 3466,
-				3270, 3465,
-				3271, 3465,
-				3271, 3437,
-				3274, 3437,
-				3274, 3424,
-				3277, 3424,
-				3277, 3420,
-				3274, 3420,
-				3274, 3411,
-				3275, 3411,
-				3275, 3410,
-				3276, 3410,
-				3276, 3409,
-				3277, 3409,
-				3277, 3408,
-				3288, 3408,
-				3288, 3391,
-				3289, 3391,
-				3289, 3385,
-				3290, 3385,
-				3290, 3378,
-				3289, 3378,
-				3289, 3377,
-				3288, 3377,
-				3288, 3376,
-				3265, 3376,
-				3265, 3380,
-				3253, 3380,
-				3253, 3382,
-				3245, 3382,
-				3245, 3380,
-				3242, 3380,
-				3242, 3382,
-				3239, 3382,
-				3239, 3381,
-				3209, 3381,
-				3209, 3382,
-				3282, 3382);
-
-		// Lumbridge
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				3201, 3257,
-				3213, 3257,
-				3213, 3264,
-				3233, 3264,
-				3233, 3257,
-				3235, 3257,
-				3235, 3241,
-				3237, 3241,
-				3237, 3237,
-				3239, 3237,
-				3239, 3231,
-				3243, 3231,
-				3243, 3220,
-				3253, 3220,
-				3253, 3217,
-				3256, 3217,
-				3256, 3212,
-				3259, 3212,
-				3259, 3190,
-				3247, 3190,
-				3247, 3191,
-				3238, 3191,
-				3238, 3195,
-				3230, 3195,
-				3230, 3201,
-				3228, 3201,
-				3228, 3202,
-				3227, 3202,
-				3227, 3205,
-				3228, 3205,
-				3228, 3207,
-				3225, 3207,
-				3225, 3206,
-				3224, 3206,
-				3224, 3205,
-				3223, 3205,
-				3223, 3204,
-				3222, 3204,
-				3222, 3203,
-				3215, 3203,
-				3215, 3202,
-				3214, 3202,
-				3214, 3201,
-				3203, 3201,
-				3203, 3202,
-				3202, 3202,
-				3202, 3203,
-				3201, 3203,
-				3201, 3217,
-				3199, 3217,
-				3199, 3220,
-				3201, 3220);
-
-		// Falador
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				2986, 3395,
-				2986, 3394,
-				2987, 3394,
-				2987, 3393,
-				2996, 3393,
-				2996, 3394,
-				3002, 3394,
-				3002, 3395,
-				3009, 3395,
-				3009, 3394,
-				3010, 3394,
-				3010, 3393,
-				3011, 3393,
-				3011, 3392,
-				3021, 3392,
-				3021, 3391,
-				3022, 3391,
-				3022, 3390,
-				3041, 3390,
-				3041, 3389,
-				3047, 3389,
-				3047, 3390,
-				3062, 3390,
-				3062, 3389,
-				3063, 3389,
-				3063, 3388,
-				3064, 3388,
-				3064, 3387,
-				3065, 3387,
-				3065, 3386,
-				3066, 3386,
-				3066, 3368,
-				3065, 3368,
-				3065, 3367,
-				3064, 3367,
-				3064, 3366,
-				3063, 3366,
-				3063, 3365,
-				3062, 3365,
-				3062, 3364,
-				3061, 3364,
-				3061, 3363,
-				3060, 3363,
-				3060, 3331,
-				3061, 3331,
-				3061, 3328,
-				3058, 3328,
-				3058, 3329,
-				3025, 3329,
-				3025, 3328,
-				3024, 3328,
-				3024, 3327,
-				3016, 3327,
-				3016, 3326,
-				3015, 3326,
-				3015, 3325,
-				3014, 3325,
-				3014, 3324,
-				3013, 3324,
-				3013, 3323,
-				3008, 3323,
-				3008, 3324,
-				3006, 3324,
-				3006, 3323,
-				3002, 3323,
-				3002, 3322,
-				3001, 3322,
-				3001, 3321,
-				3000, 3321,
-				3000, 3320,
-				2999, 3320,
-				2999, 3319,
-				2998, 3319,
-				2998, 3318,
-				2997, 3318,
-				2997, 3317,
-				2996, 3317,
-				2996, 3316,
-				2992, 3316,
-				2992, 3315,
-				2991, 3315,
-				2991, 3314,
-				2990, 3314,
-				2990, 3313,
-				2989, 3313,
-				2989, 3312,
-				2988, 3312,
-				2988, 3311,
-				2987, 3311,
-				2987, 3310,
-				2986, 3310,
-				2986, 3309,
-				2966, 3309,
-				2966, 3310,
-				2956, 3310,
-				2956, 3311,
-				2941, 3311,
-				2941, 3312,
-				2940, 3312,
-				2940, 3320,
-				2936, 3320,
-				2936, 3354,
-				2937, 3354,
-				2937, 3357,
-				2936, 3357,
-				2936, 3389,
-				2937, 3389,
-				2937, 3390,
-				2938, 3390,
-				2938, 3391,
-				2939, 3391,
-				2939, 3392,
-				2940, 3392,
-				2940, 3393,
-				2943, 3393,
-				2943, 3394,
-				2944, 3394,
-				2944, 3395,
-				2950, 3395,
-				2950, 3394,
-				2956, 3394,
-				2956, 3395);
-
-		// Port Phasmatys
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				3650, 3456,
-				3650, 3472,
-				3651, 3472,
-				3651, 3473,
-				3652, 3473,
-				3652, 3474,
-				3653, 3474,
-				3653, 3507,
-				3654, 3507,
-				3654, 3508,
-				3668, 3508,
-				3668, 3509,
-				3669, 3509,
-				3669, 3510,
-				3670, 3510,
-				3670, 3511,
-				3671, 3511,
-				3671, 3512,
-				3672, 3512,
-				3672, 3513,
-				3673, 3513,
-				3673, 3514,
-				3674, 3514,
-				3674, 3515,
-				3675, 3515,
-				3675, 3516,
-				3676, 3516,
-				3676, 3517,
-				3687, 3517,
-				3687, 3494,
-				3690, 3494,
-				3690, 3493,
-				3696, 3493,
-				3696, 3482,
-				3699, 3482,
-				3699, 3481,
-				3712, 3481,
-				3712, 3456);
-
-		// Sophanem
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				3274, 2752,
-				3274, 2784,
-				3277, 2784,
-				3277, 2786,
-				3274, 2786,
-				3274, 2789,
-				3272, 2789,
-				3272, 2810,
-				3322, 2810,
-				3322, 2752);
-
-		// Ardougne
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				2560, 3256,
-				2560, 3264,
-				2559, 3264,
-				2559, 3328,
-				2560, 3328,
-				2560, 3339,
-				2561, 3339,
-				2561, 3340,
-				2562, 3340,
-				2562, 3341,
-				2563, 3341,
-				2563, 3342,
-				2616, 3342,
-				2616, 3341,
-				2617, 3341,
-				2617, 3340,
-				2669, 3340,
-				2669, 3339,
-				2670, 3339,
-				2670, 3338,
-				2671, 3338,
-				2671, 3337,
-				2672, 3337,
-				2672, 3336,
-				2673, 3336,
-				2673, 3335,
-				2674, 3335,
-				2674, 3334,
-				2683, 3334,
-				2683, 3333,
-				2684, 3333,
-				2684, 3332,
-				2685, 3332,
-				2685, 3331,
-				2686, 3331,
-				2686, 3330,
-				2687, 3330,
-				2687, 3329,
-				2688, 3329,
-				2688, 3264,
-				2638, 3264,
-				2638, 3263,
-				2625, 3263,
-				2625, 3264,
-				2611, 3264,
-				2611, 3257,
-				2602, 3257,
-				2602, 3264,
-				2587, 3264,
-				2587, 3263,
-				2586, 3263,
-				2586, 3262,
-				2584, 3262,
-				2584, 3261,
-				2583, 3261,
-				2583, 3260,
-				2582, 3260,
-				2582, 3259,
-				2581, 3259,
-				2581, 3258,
-				2572, 3258,
-				2572, 3260,
-				2571, 3260,
-				2571, 3261,
-				2566, 3261,
-				2566, 3260,
-				2565, 3260,
-				2565, 3259,
-				2564, 3259,
-				2564, 3256);
-
-		// Yanille
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				2613, 3103,
-				2614, 3103,
-				2614, 3102,
-				2615, 3102,
-				2615, 3101,
-				2616, 3101,
-				2616, 3100,
-				2617, 3100,
-				2617, 3099,
-				2618, 3099,
-				2618, 3098,
-				2619, 3098,
-				2619, 3097,
-				2620, 3097,
-				2620, 3075,
-				2590, 3075,
-				2590, 3074,
-				2589, 3074,
-				2589, 3073,
-				2584, 3073,
-				2584, 3074,
-				2583, 3074,
-				2583, 3075,
-				2543, 3075,
-				2543, 3076,
-				2542, 3076,
-				2542, 3077,
-				2539, 3077,
-				2539, 3107,
-				2542, 3107,
-				2542, 3108,
-				2543, 3108,
-				2543, 3109,
-				2608, 3109,
-				2608, 3108,
-				2609, 3108,
-				2609, 3107,
-				2610, 3107,
-				2610, 3106,
-				2611, 3106,
-				2611, 3105,
-				2612, 3105,
-				2612, 3104,
-				2613, 3104);
-
-		// Gnome stronghold
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				2495, 3439,
-				2494, 3439,
-				2494, 3432,
-				2495, 3432,
-				2495, 3431,
-				2496, 3431,
-				2496, 3430,
-				2497, 3430,
-				2497, 3429,
-				2498, 3429,
-				2498, 3417,
-				2497, 3417,
-				2497, 3416,
-				2496, 3416,
-				2496, 3412,
-				2495, 3412,
-				2495, 3408,
-				2494, 3408,
-				2494, 3404,
-				2495, 3404,
-				2495, 3403,
-				2496, 3403,
-				2496, 3402,
-				2497, 3402,
-				2497, 3401,
-				2498, 3401,
-				2498, 3400,
-				2499, 3400,
-				2499, 3399,
-				2500, 3399,
-				2500, 3398,
-				2501, 3398,
-				2501, 3397,
-				2502, 3397,
-				2502, 3396,
-				2506, 3396,
-				2506, 3391,
-				2502, 3391,
-				2502, 3390,
-				2492, 3390,
-				2492, 3391,
-				2489, 3391,
-				2489, 3390,
-				2488, 3390,
-				2488, 3389,
-				2485, 3389,
-				2485, 3390,
-				2482, 3390,
-				2482, 3389,
-				2476, 3389,
-				2476, 3390,
-				2471, 3390,
-				2471, 3391,
-				2468, 3391,
-				2468, 3390,
-				2467, 3390,
-				2467, 3389,
-				2466, 3389,
-				2466, 3385,
-				2465, 3385,
-				2465, 3384,
-				2458, 3384,
-				2458, 3385,
-				2457, 3385,
-				2457, 3389,
-				2456, 3389,
-				2456, 3390,
-				2455, 3390,
-				2455, 3391,
-				2450, 3391,
-				2450, 3390,
-				2446, 3390,
-				2446, 3391,
-				2443, 3391,
-				2443, 3390,
-				2442, 3390,
-				2442, 3389,
-				2440, 3389,
-				2440, 3388,
-				2434, 3388,
-				2434, 3389,
-				2433, 3389,
-				2433, 3390,
-				2432, 3390,
-				2432, 3391,
-				2428, 3391,
-				2428, 3392,
-				2427, 3392,
-				2427, 3393,
-				2420, 3393,
-				2420, 3394,
-				2419, 3394,
-				2419, 3395,
-				2418, 3395,
-				2418, 3396,
-				2417, 3396,
-				2417, 3397,
-				2416, 3397,
-				2416, 3399,
-				2415, 3399,
-				2415, 3400,
-				2414, 3400,
-				2414, 3408,
-				2413, 3408,
-				2413, 3409,
-				2412, 3409,
-				2412, 3410,
-				2411, 3410,
-				2411, 3411,
-				2410, 3411,
-				2410, 3412,
-				2387, 3412,
-				2387, 3407,
-				2383, 3407,
-				2383, 3408,
-				2380, 3408,
-				2380, 3409,
-				2379, 3409,
-				2379, 3410,
-				2377, 3410,
-				2377, 3411,
-				2376, 3411,
-				2376, 3413,
-				2375, 3413,
-				2375, 3417,
-				2374, 3417,
-				2374, 3418,
-				2373, 3418,
-				2373, 3419,
-				2372, 3419,
-				2372, 3420,
-				2371, 3420,
-				2371, 3421,
-				2370, 3421,
-				2370, 3422,
-				2369, 3422,
-				2369, 3433,
-				2370, 3433,
-				2370, 3434,
-				2371, 3434,
-				2371, 3444,
-				2372, 3444,
-				2372, 3445,
-				2373, 3445,
-				2373, 3446,
-				2374, 3446,
-				2374, 3447,
-				2375, 3447,
-				2375, 3459,
-				2376, 3459,
-				2376, 3460,
-				2377, 3460,
-				2377, 3461,
-				2378, 3461,
-				2378, 3462,
-				2379, 3462,
-				2379, 3463,
-				2380, 3463,
-				2380, 3464,
-				2381, 3464,
-				2381, 3476,
-				2379, 3476,
-				2379, 3477,
-				2378, 3477,
-				2378, 3478,
-				2377, 3478,
-				2377, 3485,
-				2376, 3485,
-				2376, 3486,
-				2375, 3486,
-				2375, 3499,
-				2376, 3499,
-				2376, 3500,
-				2377, 3500,
-				2377, 3507,
-				2378, 3507,
-				2378, 3508,
-				2379, 3508,
-				2379, 3509,
-				2380, 3509,
-				2380, 3521,
-				2382, 3521,
-				2382, 3522,
-				2384, 3522,
-				2384, 3523,
-				2393, 3523,
-				2393, 3524,
-				2399, 3524,
-				2399, 3525,
-				2404, 3525,
-				2404, 3524,
-				2405, 3524,
-				2405, 3523,
-				2407, 3523,
-				2407, 3522,
-				2415, 3522,
-				2415, 3521,
-				2425, 3521,
-				2425, 3522,
-				2427, 3522,
-				2427, 3523,
-				2430, 3523,
-				2430, 3522,
-				2431, 3522,
-				2431, 3521,
-				2432, 3521,
-				2432, 3520,
-				2448, 3520,
-				2448, 3517,
-				2454, 3517,
-				2454, 3516,
-				2455, 3516,
-				2455, 3515,
-				2456, 3515,
-				2456, 3514,
-				2457, 3514,
-				2457, 3513,
-				2460, 3513,
-				2460, 3512,
-				2461, 3512,
-				2461, 3511,
-				2465, 3511,
-				2465, 3510,
-				2468, 3510,
-				2468, 3511,
-				2472, 3511,
-				2472, 3512,
-				2473, 3512,
-				2473, 3513,
-				2475, 3513,
-				2475, 3514,
-				2476, 3514,
-				2476, 3515,
-				2477, 3515,
-				2477, 3516,
-				2478, 3516,
-				2478, 3517,
-				2483, 3517,
-				2483, 3516,
-				2487, 3516,
-				2487, 3515,
-				2488, 3515,
-				2488, 3512,
-				2487, 3512,
-				2487, 3509,
-				2488, 3509,
-				2488, 3508,
-				2489, 3508,
-				2489, 3507,
-				2491, 3507,
-				2491, 3506,
-				2492, 3506,
-				2492, 3505,
-				2493, 3505,
-				2493, 3499,
-				2492, 3499,
-				2492, 3498,
-				2491, 3498,
-				2491, 3497,
-				2490, 3497,
-				2490, 3495,
-				2491, 3495,
-				2491, 3494,
-				2492, 3494,
-				2492, 3493,
-				2493, 3493,
-				2493, 3485,
-				2490, 3485,
-				2490, 3484,
-				2489, 3484,
-				2489, 3483,
-				2488, 3483,
-				2488, 3482,
-				2487, 3482,
-				2487, 3481,
-				2486, 3481,
-				2486, 3474,
-				2488, 3474,
-				2488, 3471,
-				2489, 3471,
-				2489, 3470,
-				2490, 3470,
-				2490, 3460,
-				2491, 3460,
-				2491, 3456,
-				2496, 3456,
-				2496, 3440,
-				2495, 3440);
-
-		// Rellekka
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				2620, 3682,
-				2624, 3682,
-				2624, 3683,
-				2625, 3683,
-				2625, 3687,
-				2629, 3687,
-				2629, 3686,
-				2630, 3686,
-				2630, 3685,
-				2632, 3685,
-				2632, 3686,
-				2636, 3686,
-				2636, 3692,
-				2645, 3692,
-				2645, 3695,
-				2647, 3695,
-				2647, 3696,
-				2649, 3696,
-				2649, 3702,
-				2650, 3702,
-				2650, 3703,
-				2651, 3703,
-				2651, 3704,
-				2652, 3704,
-				2652, 3711,
-				2653, 3711,
-				2653, 3712,
-				2691, 3712,
-				2691, 3709,
-				2692, 3709,
-				2692, 3707,
-				2693, 3707,
-				2693, 3703,
-				2692, 3703,
-				2692, 3701,
-				2691, 3701,
-				2691, 3699,
-				2690, 3699,
-				2690, 3695,
-				2691, 3695,
-				2691, 3693,
-				2692, 3693,
-				2692, 3691,
-				2693, 3691,
-				2693, 3685,
-				2692, 3685,
-				2692, 3683,
-				2691, 3683,
-				2691, 3681,
-				2690, 3681,
-				2690, 3680,
-				2689, 3680,
-				2689, 3672,
-				2690, 3672,
-				2690, 3671,
-				2691, 3671,
-				2691, 3666,
-				2690, 3666,
-				2690, 3664,
-				2689, 3664,
-				2689, 3660,
-				2690, 3660,
-				2690, 3658,
-				2691, 3658,
-				2691, 3656,
-				2692, 3656,
-				2692, 3654,
-				2693, 3654,
-				2693, 3651,
-				2692, 3651,
-				2692, 3649,
-				2690, 3649,
-				2690, 3648,
-				2688, 3648,
-				2688, 3647,
-				2686, 3647,
-				2686, 3646,
-				2673, 3646,
-				2673, 3645,
-				2636, 3645,
-				2636, 3647,
-				2627, 3647,
-				2627, 3648,
-				2625, 3648,
-				2625, 3649,
-				2624, 3649,
-				2624, 3650,
-				2622, 3650,
-				2622, 3651,
-				2620, 3651,
-				2620, 3652,
-				2618, 3652,
-				2618, 3653,
-				2616, 3653,
-				2616, 3654,
-				2609, 3654,
-				2609, 3655,
-				2607, 3655,
-				2607, 3656,
-				2603, 3656,
-				2603, 3657,
-				2602, 3657,
-				2602, 3658,
-				2601, 3658,
-				2601, 3663,
-				2602, 3663,
-				2602, 3664,
-				2603, 3664,
-				2603, 3665,
-				2604, 3665,
-				2604, 3666,
-				2605, 3666,
-				2605, 3667,
-				2606, 3667,
-				2606, 3671,
-				2609, 3671,
-				2609, 3672,
-				2610, 3672,
-				2610, 3673,
-				2611, 3673,
-				2611, 3675,
-				2612, 3675,
-				2612, 3676,
-				2614, 3676,
-				2614, 3677,
-				2616, 3677,
-				2616, 3679,
-				2618, 3679,
-				2618, 3681,
-				2620, 3681);
-
-		// Jatizso
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				2407, 3797,
-				2407, 3793,
-				2399, 3793,
-				2399, 3792,
-				2391, 3792,
-				2391, 3791,
-				2386, 3791,
-				2386, 3796,
-				2388, 3796,
-				2388, 3802,
-				2386, 3802,
-				2386, 3807,
-				2388, 3807,
-				2388, 3809,
-				2402, 3809,
-				2402, 3819,
-				2406, 3819,
-				2406, 3824,
-				2408, 3824,
-				2408, 3826,
-				2413, 3826,
-				2413, 3824,
-				2419, 3824,
-				2419, 3826,
-				2424, 3826,
-				2424, 3821,
-				2423, 3821,
-				2423, 3798,
-				2422, 3798,
-				2422, 3797);
-
-		// Neitiznot
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				2329, 3812,
-				2333, 3812,
-				2333, 3813,
-				2334, 3813,
-				2334, 3814,
-				2335, 3814,
-				2335, 3815,
-				2338, 3815,
-				2338, 3816,
-				2339, 3816,
-				2339, 3817,
-				2368, 3817,
-				2368, 3776,
-				2352, 3776,
-				2352, 3796,
-				2344, 3796,
-				2344, 3795,
-				2331, 3795,
-				2331, 3797,
-				2330, 3797,
-				2330, 3798,
-				2329, 3798);
-
-		// Pest control
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				2624, 2688,
-				2688, 2688,
-				2688, 2624,
-				2624, 2624);
-
-		// Tutorial island
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				3052, 3135,
-				3156, 3135,
-				3156, 3057,
-				3052, 3057);
-
-		// Camelot bank
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				2724, 3487,
-				2724, 3490,
-				2721, 3490,
-				2721, 3494,
-				2719, 3494,
-				2719, 3497,
-				2721, 3497,
-				2721, 3498,
-				2731, 3498,
-				2731, 3490,
-				2728, 3490,
-				2728, 3487);
-
-		// Catherby bank
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				2806, 3438,
-				2806, 3446,
-				2813, 3446,
-				2813, 3438);
-
-		// Kourend castle
-		addPolygonOnPlane(DEADMAN_SAFE_ZONES, 0,
-				1627, 3658,
-				1620, 3658,
-				1620, 3657,
-				1619, 3657,
-				1619, 3656,
-				1614, 3656,
-				1614, 3657,
-				1613, 3657,
-				1613, 3661,
-				1612, 3661,
-				1612, 3662,
-				1611, 3662,
-				1611, 3663,
-				1600, 3663,
-				1600, 3662,
-				1599, 3662,
-				1599, 3661,
-				1594, 3661,
-				1594, 3662,
-				1593, 3662,
-				1593, 3685,
-				1594, 3685,
-				1594, 3686,
-				1599, 3686,
-				1599, 3685,
-				1600, 3685,
-				1600, 3684,
-				1611, 3684,
-				1611, 3685,
-				1612, 3685,
-				1612, 3686,
-				1613, 3686,
-				1613, 3690,
-				1614, 3690,
-				1614, 3691,
-				1619, 3691,
-				1619, 3690,
-				1620, 3690,
-				1620, 3689,
-				1630, 3689,
-				1630, 3686,
-				1620, 3686,
-				1620, 3685,
-				1619, 3685,
-				1619, 3683,
-				1620, 3683,
-				1620, 3682,
-				1621, 3682,
-				1621, 3681,
-				1622, 3681,
-				1622, 3680,
-				1623, 3680,
-				1623, 3679,
-				1624, 3679,
-				1624, 3668,
-				1623, 3668,
-				1623, 3667,
-				1622, 3667,
-				1622, 3666,
-				1621, 3666,
-				1621, 3665,
-				1620, 3665,
-				1620, 3664,
-				1619, 3664,
-				1619, 3662,
-				1620, 3662,
-				1620, 3661,
-				1627, 3661);
-	}
-
-	private static void definePvpSafeZones()
-	{
-		// Grand Exchange
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3159, 3473,
-				3159, 3474,
-				3157, 3474,
-				3157, 3475,
-				3155, 3475,
-				3155, 3476,
-				3153, 3476,
-				3153, 3477,
-				3152, 3477,
-				3152, 3478,
-				3151, 3478,
-				3151, 3480,
-				3150, 3480,
-				3150, 3482,
-				3149, 3482,
-				3149, 3484,
-				3148, 3484,
-				3148, 3496,
-				3149, 3496,
-				3149, 3498,
-				3150, 3498,
-				3150, 3500,
-				3151, 3500,
-				3151, 3502,
-				3152, 3502,
-				3152, 3503,
-				3153, 3503,
-				3153, 3504,
-				3155, 3504,
-				3155, 3505,
-				3157, 3505,
-				3157, 3506,
-				3159, 3506,
-				3159, 3507,
-				3171, 3507,
-				3171, 3506,
-				3173, 3506,
-				3173, 3505,
-				3175, 3505,
-				3175, 3504,
-				3177, 3504,
-				3177, 3503,
-				3178, 3503,
-				3178, 3502,
-				3179, 3502,
-				3179, 3500,
-				3180, 3500,
-				3180, 3498,
-				3181, 3498,
-				3181, 3496,
-				3182, 3496,
-				3182, 3484,
-				3181, 3484,
-				3181, 3482,
-				3180, 3482,
-				3180, 3480,
-				3179, 3480,
-				3179, 3478,
-				3178, 3478,
-				3178, 3477,
-				3177, 3477,
-				3177, 3476,
-				3175, 3476,
-				3175, 3475,
-				3173, 3475,
-				3173, 3474,
-				3171, 3474,
-				3171, 3473);
-
-		// Edgeville
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3091, 3488,
-				3091, 3493,
-				3090, 3493,
-				3090, 3498,
-				3091, 3498,
-				3091, 3500,
-				3099, 3500,
-				3099, 3488);
-
-		// Falador west bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2943, 3368,
-				2943, 3374,
-				2948, 3374,
-				2948, 3370,
-				2950, 3370,
-				2950, 3366,
-				2949, 3366,
-				2949, 3359,
-				2945, 3359,
-				2945, 3362,
-				2946, 3362,
-				2946, 3366,
-				2945, 3366,
-				2945, 3368);
-
-		// Falador east bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3009, 3353,
-				3009, 3359,
-				3019, 3359,
-				3019, 3357,
-				3022, 3357,
-				3022, 3353);
-
-		// Falador castle
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2964, 3354,
-				2966, 3354,
-				2966, 3352,
-				2967, 3352,
-				2967, 3349,
-				2976, 3349,
-				2976, 3348,
-				2977, 3348,
-				2977, 3347,
-				2981, 3347,
-				2981, 3343,
-				2982, 3343,
-				2982, 3339,
-				2981, 3339,
-				2981, 3337,
-				2967, 3337,
-				2967, 3330,
-				2963, 3330,
-				2963, 3331,
-				2962, 3331,
-				2962, 3332,
-				2961, 3332,
-				2961, 3334,
-				2964, 3334,
-				2964, 3335,
-				2965, 3335,
-				2965, 3343,
-				2964, 3343,
-				2964, 3344,
-				2961, 3344,
-				2961, 3350,
-				2963, 3350,
-				2963, 3352,
-				2964, 3352);
-
-		// Varrock east bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3250, 3425,
-				3258, 3425,
-				3258, 3416,
-				3250, 3416);
-
-		// Varrock west bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3180, 3433,
-				3180, 3448,
-				3191, 3448,
-				3191, 3433);
-
-		// Port Phasmatys
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3686, 3472,
-				3700, 3472,
-				3700, 3461,
-				3686, 3461);
-
-		// Yanille bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2609, 3088,
-				2609, 3098,
-				2617, 3098,
-				2617, 3088);
-
-		// Ardougne east bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2649, 3280,
-				2649, 3288,
-				2659, 3288,
-				2659, 3280);
-
-		// Ardougne west bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2612, 3330,
-				2612, 3336,
-				2615, 3336,
-				2615, 3335,
-				2619, 3335,
-				2619, 3336,
-				2622, 3336,
-				2622, 3330);
-
-		// Fishing guild bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2593, 3413,
-				2588, 3413,
-				2588, 3418,
-				2583, 3418,
-				2583, 3423,
-				2590, 3423,
-				2590, 3420,
-				2593, 3420);
-
-		// Gnome stronghold bank near slayer cave (2nd floor)
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 1,
-				2444, 3431,
-				2444, 3435,
-				2448, 3435,
-				2448, 3431,
-				2447, 3431,
-				2447, 3428,
-				2449, 3428,
-				2449, 3422,
-				2447, 3422,
-				2447, 3419,
-				2448, 3419,
-				2448, 3415,
-				2444, 3415,
-				2444, 3419,
-				2445, 3419,
-				2445, 3422,
-				2443, 3422,
-				2443, 3428,
-				2445, 3428,
-				2445, 3431);
-
-		// Gnome stronghold bank in grand tree
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 1,
-				2456, 3488,
-				2452, 3488,
-				2452, 3486,
-				2450, 3486,
-				2450, 3483,
-				2451, 3483,
-				2451, 3478,
-				2448, 3478,
-				2448, 3483,
-				2449, 3483,
-				2449, 3486,
-				2447, 3486,
-				2447, 3488,
-				2443, 3488,
-				2443, 3487,
-				2438, 3487,
-				2438, 3490,
-				2443, 3490,
-				2443, 3489,
-				2447, 3489,
-				2447, 3491,
-				2449, 3491,
-				2449, 3494,
-				2448, 3494,
-				2448, 3496,
-				2451, 3496,
-				2451, 3494,
-				2450, 3494,
-				2450, 3491,
-				2452, 3491,
-				2452, 3489,
-				2456, 3489);
-
-		// Al kharid bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3265, 3161,
-				3265, 3174,
-				3273, 3174,
-				3273, 3161);
-
-		// Shantay pass bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3308, 3119,
-				3308, 3125,
-				3310, 3125,
-				3310, 3119);
-
-		// Nardah bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3431, 2891,
-				3431, 2889,
-				3427, 2889,
-				3427, 2887,
-				3424, 2887,
-				3424, 2895,
-				3431, 2895,
-				3431, 2893,
-				3432, 2893,
-				3432, 2891);
-
-		// Sophanem bank
-		addPolygonTo(PVP_WORLD_SAFE_ZONES,
-				2807, 5158,
-				2792, 5158,
-				2792, 5175,
-				2807, 5175);
-
-		// Canifis bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3509, 3474,
-				3509, 3478,
-				3508, 3478,
-				3508, 3483,
-				3509, 3483,
-				3509, 3484,
-				3517, 3484,
-				3517, 3477,
-				3516, 3477,
-				3516, 3476,
-				3513, 3476,
-				3513, 3474);
-
-		// Lumbridge castle outside
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3216, 3209,
-				3216, 3210,
-				3217, 3210,
-				3217, 3228,
-				3216, 3228,
-				3216, 3229,
-				3227, 3229,
-				3227, 3221,
-				3230, 3221,
-				3230, 3217,
-				3227, 3217,
-				3227, 3209);
-
-		// Lumbridge bank upstairs
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 2,
-				3211, 3223,
-				3211, 3215,
-				3207, 3215,
-				3207, 3223);
-
-		// Draynor bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3098, 3240,
-				3088, 3240,
-				3088, 3247,
-				3098, 3247);
-
-		// Pest control bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2665, 2656,
-				2670, 2656,
-				2670, 2651,
-				2665, 2651);
-
-		// Shilo village bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2843, 2957,
-				2846, 2957,
-				2846, 2956,
-				2849, 2956,
-				2849, 2957,
-				2850, 2957,
-				2850, 2958,
-				2855, 2958,
-				2855, 2957,
-				2856, 2957,
-				2856, 2956,
-				2858, 2956,
-				2858, 2957,
-				2862, 2957,
-				2862, 2952,
-				2858, 2952,
-				2858, 2953,
-				2856, 2953,
-				2856, 2952,
-				2855, 2952,
-				2855, 2951,
-				2850, 2951,
-				2850, 2952,
-				2849, 2952,
-				2849, 2953,
-				2847, 2953,
-				2847, 2952,
-				2843, 2952);
-
-		// Legends guild bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 2,
-				2731, 3374,
-				2731, 3383,
-				2734, 3383,
-				2734, 3374);
-
-		// Legends guild middle floor
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 1,
-				2724, 3374,
-				2724, 3383,
-				2734, 3383,
-				2734, 3382,
-				2736, 3382,
-				2736, 3375,
-				2734, 3375,
-				2734, 3374);
-
-		// Warriors guild bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2843, 3537,
-				2843, 3540,
-				2841, 3540,
-				2841, 3546,
-				2849, 3546,
-				2849, 3537,
-				2847, 3537,
-				2847, 3536,
-				2846, 3536,
-				2846, 3537);
-
-		// Camelot bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2724, 3487,
-				2724, 3490,
-				2721, 3490,
-				2721, 3494,
-				2719, 3494,
-				2719, 3497,
-				2721, 3497,
-				2721, 3498,
-				2731, 3498,
-				2731, 3490,
-				2728, 3490,
-				2728, 3487);
-
-		// Camelot respawn point
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2761, 3483,
-				2761, 3476,
-				2755, 3476,
-				2755, 3483);
-
-		// Catherby bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2806, 3438,
-				2806, 3446,
-				2813, 3446,
-				2813, 3438);
-
-		// Barbarian outpost bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2536, 3572,
-				2536, 3575,
-				2538, 3575,
-				2538, 3572);
-
-		// Piscatoris bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2327, 3686,
-				2327, 3694,
-				2333, 3694,
-				2333, 3686);
-
-		// Lletya bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2350, 3161,
-				2350, 3165,
-				2351, 3165,
-				2351, 3167,
-				2357, 3167,
-				2357, 3165,
-				2356, 3165,
-				2356, 3164,
-				2355, 3164,
-				2355, 3161);
-
-		// Castle wars bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2446, 3087,
-				2445, 3087,
-				2445, 3085,
-				2447, 3085,
-				2447, 3081,
-				2443, 3081,
-				2443, 3082,
-				2439, 3082,
-				2439, 3081,
-				2435, 3081,
-				2435, 3099,
-				2439, 3099,
-				2439, 3098,
-				2443, 3098,
-				2443, 3099,
-				2447, 3099,
-				2447, 3095,
-				2445, 3095,
-				2445, 3093,
-				2446, 3093);
-
-		// Duel arena bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3380, 3267,
-				3380, 3273,
-				3381, 3273,
-				3381, 3274,
-				3385, 3274,
-				3385, 3267);
-
-		// Clan wars bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3375, 3165,
-				3361, 3165,
-				3361, 3173,
-				3375, 3173);
-
-		// Lumbridge cellar bank
-		addPolygonTo(PVP_WORLD_SAFE_ZONES,
-				3218, 9622,
-				3218, 9624,
-				3220, 9624,
-				3220, 9622);
-
-		// Dorgesh kaan bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2709, 5348,
-				2707, 5348,
-				2707, 5345,
-				2701, 5345,
-				2701, 5347,
-				2697, 5347,
-				2697, 5353,
-				2701, 5353,
-				2701, 5355,
-				2707, 5355,
-				2707, 5350,
-				2709, 5350);
-
-		// Keldagrim bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2842, 10204,
-				2834, 10204,
-				2834, 10216,
-				2842, 10216);
-
-		// Tzhaar bank
-		addPolygonTo(PVP_WORLD_SAFE_ZONES,
-				2438, 5176,
-				2438, 5180,
-				2441, 5180,
-				2441, 5182,
-				2449, 5182,
-				2449, 5181,
-				2450, 5181,
-				2450, 5180,
-				2452, 5180,
-				2452, 5175,
-				2441, 5175,
-				2441, 5176);
-
-		// Inferno bank
-		addPolygonTo(PVP_WORLD_SAFE_ZONES,
-				2542, 5135,
-				2542, 5139,
-				2539, 5139,
-				2539, 5140,
-				2538, 5140,
-				2538, 5141,
-				2537, 5141,
-				2537, 5144,
-				2541, 5144,
-				2541, 5145,
-				2543, 5145,
-				2543, 5144,
-				2544, 5144,
-				2544, 5142,
-				2545, 5142,
-				2545, 5135);
-
-		// Port Khazard bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2661, 3160,
-				2661, 3163,
-				2666, 3163,
-				2666, 3160);
-
-		// Corsair cove bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2569, 2863,
-				2569, 2868,
-				2572, 2868,
-				2572, 2863);
-
-		// Burgh de Rott bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3495, 3210,
-				3495, 3214,
-				3501, 3214,
-				3501, 3210);
-
-		// Edgeville respawn point
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3092, 3468,
-				3092, 3474,
-				3098, 3474,
-				3098, 3468);
-
-		// Mage bank
-		addPolygonTo(PVP_WORLD_SAFE_ZONES,
-				2529, 4711,
-				2529, 4724,
-				2548, 4724,
-				2548, 4711);
-
-		// Lunar bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2097, 3917,
-				2097, 3922,
-				2105, 3922,
-				2105, 3917);
-
-		// Jatizso bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2414, 3801,
-				2414, 3804,
-				2420, 3804,
-				2420, 3801);
-
-		// Neitiznot bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2334, 3805,
-				2334, 3809,
-				2340, 3809,
-				2340, 3805);
-
-		// Hosidius bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1671, 3558,
-				1671, 3577,
-				1682, 3577,
-				1682, 3558);
-
-		// Woodcutting guild bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1589, 3475,
-				1589, 3481,
-				1594, 3481,
-				1594, 3475);
-
-		// Land's End bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1508, 3415,
-				1508, 3424,
-				1514, 3424,
-				1514, 3415);
-
-		// Chambers of Xeric bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1252, 3570,
-				1252, 3574,
-				1257, 3574,
-				1257, 3570);
-
-		// Arceuus bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1621, 3736,
-				1621, 3754,
-				1627, 3754,
-				1627, 3751,
-				1633, 3751,
-				1633, 3754,
-				1639, 3754,
-				1639, 3736);
-
-		// Piscarilius bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1794, 3784,
-				1794, 3794,
-				1812, 3794,
-				1812, 3784);
-
-		// Lovakengj bank southeast
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1518, 3735,
-				1518, 3744,
-				1535, 3744,
-				1535, 3735);
-
-		// Lovakengj bank west
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1433, 3820,
-				1433, 3837,
-				1442, 3837,
-				1442, 3820);
-
-		// Lovakengj sulphur mine bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1452, 3855,
-				1452, 3860,
-				1455, 3860,
-				1455, 3855);
-
-		// Blast mine bank southeast
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1500, 3856,
-				1500, 3858,
-				1503, 3858,
-				1503, 3856);
-
-		// Wintertodt bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1638, 3942,
-				1638, 3947,
-				1642, 3947,
-				1642, 3942);
-
-		// Shayzien bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1495, 3612,
-				1495, 3622,
-				1515, 3622,
-				1515, 3612);
-
-		// Hosidius grape farm bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1804, 3571,
-				1804, 3572,
-				1808, 3572,
-				1808, 3571);
-
-		// Hosidius cooking bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				1652, 3605,
-				1652, 3615,
-				1661, 3615,
-				1661, 3605);
-
-		// Etceteria bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				2618, 3893,
-				2618, 3897,
-				2622, 3897,
-				2622, 3893);
-
-		// Mining guild expanded area
-		addPolygonTo(PVP_WORLD_SAFE_ZONES,
-				3018, 9733,
-				3021, 9733,
-				3021, 9729,
-				3022, 9729,
-				3022, 9728,
-				3023, 9728,
-				3023, 9727,
-				3025, 9727,
-				3025, 9726,
-				3026, 9726,
-				3026, 9725,
-				3030, 9725,
-				3030, 9726,
-				3032, 9726,
-				3032, 9727,
-				3035, 9727,
-				3035, 9726,
-				3038, 9726,
-				3038, 9727,
-				3041, 9727,
-				3041, 9728,
-				3042, 9728,
-				3042, 9730,
-				3045, 9730,
-				3045, 9727,
-				3047, 9727,
-				3047, 9726,
-				3048, 9726,
-				3048, 9724,
-				3052, 9724,
-				3052, 9725,
-				3053, 9725,
-				3053, 9726,
-				3055, 9726,
-				3055, 9725,
-				3056, 9725,
-				3056, 9723,
-				3057, 9723,
-				3057, 9720,
-				3056, 9720,
-				3056, 9719,
-				3054, 9719,
-				3054, 9718,
-				3052, 9718,
-				3052, 9717,
-				3050, 9717,
-				3050, 9718,
-				3045, 9718,
-				3045, 9716,
-				3044, 9716,
-				3044, 9715,
-				3041, 9715,
-				3041, 9714,
-				3039, 9714,
-				3039, 9713,
-				3037, 9713,
-				3037, 9714,
-				3036, 9714,
-				3036, 9715,
-				3034, 9715,
-				3034, 9716,
-				3029, 9716,
-				3029, 9715,
-				3028, 9715,
-				3028, 9714,
-				3026, 9714,
-				3026, 9709,
-				3027, 9709,
-				3027, 9708,
-				3028, 9708,
-				3028, 9705,
-				3029, 9705,
-				3029, 9701,
-				3028, 9701,
-				3028, 9700,
-				3027, 9700,
-				3027, 9699,
-				3023, 9699,
-				3023, 9700,
-				3019, 9700,
-				3019, 9701,
-				3018, 9701,
-				3018, 9705,
-				3019, 9705,
-				3019, 9707,
-				3020, 9707,
-				3020, 9708,
-				3021, 9708,
-				3021, 9709,
-				3022, 9709,
-				3022, 9713,
-				3021, 9713,
-				3021, 9714,
-				3019, 9714,
-				3019, 9715,
-				3018, 9715,
-				3018, 9717,
-				3015, 9717,
-				3015, 9716,
-				3013, 9716,
-				3013, 9717,
-				3012, 9717,
-				3012, 9720,
-				3013, 9720,
-				3013, 9721,
-				3015, 9721,
-				3015, 9723,
-				3016, 9723,
-				3016, 9727,
-				3017, 9727,
-				3017, 9730,
-				3018, 9730);
-
-		// Motherlode mine bank
-		addPolygonTo(PVP_WORLD_SAFE_ZONES,
-				3760, 5671,
-				3760, 5668,
-				3761, 5668,
-				3761, 5665,
-				3760, 5665,
-				3760, 5663,
-				3758, 5663,
-				3758, 5671);
-
-		// Mos Le'Harmless bank
-		addPolygonOnPlane(PVP_WORLD_SAFE_ZONES, 0,
-				3679, 2980,
-				3679, 2985,
-				3681, 2985,
-				3681, 2984,
-				3682, 2984,
-				3682, 2985,
-				3684, 2985,
-				3684, 2980,
-				3682, 2980,
-				3682, 2981,
-				3681, 2981,
-				3681, 2980);
-
-		// Zanaris bank
-		addPolygonTo(PVP_WORLD_SAFE_ZONES,
-				2388, 4454,
-				2380, 4454,
-				2380, 4463,
-				2388, 4463);
-
-		// Woodcutting guild bank underground
-		addPolygonTo(PVP_WORLD_SAFE_ZONES,
-				1550, 9872,
-				1550, 9874,
-				1553, 9874,
-				1553, 9872);
-	}
-
-	private static void defineWilderness()
-	{
-		// Above ground
-		addPolygonTo(WILDERNESS,
-				2944, 3524,
-				3392, 3524,
-				3392, 3971,
-				2944, 3971);
-
-		// Underground
-		addPolygonTo(WILDERNESS,
-				2944, 9918,
-				2944, 10360,
-				3264, 10360,
-				3264, 9918);
-	}
-
-	private static void addPolygonTo(List<Shape>[] shapes, int... coords)
-	{
-		Polygon poly = new Polygon();
-		for (int i = 0; i < coords.length; i += 2)
-		{
-			poly.addPoint(coords[i], coords[i + 1]);
-		}
-		for (List<Shape> shape : shapes)
-		{
-			shape.add(poly);
-		}
-	}
-
-	private static void addPolygonOnPlane(List<Shape>[] shapes, int plane, int... coords)
-	{
-		Polygon poly = new Polygon();
-		for (int i = 0; i < coords.length; i += 2)
-		{
-			poly.addPoint(coords[i], coords[i + 1]);
-		}
-		shapes[plane].add(poly);
-	}
-}
