+++ runelite-client/src/main/java/net/runelite/client/plugins/spelleffecttimers/SpellEffectTimersConfig.java
@@ -1,3 +1,27 @@
+/*
+ * Copyright (c) 2019, Siraz, Jajack
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
 package net.runelite.client.plugins.spelleffecttimers;
 
 import net.runelite.client.config.Config;
@@ -36,7 +60,7 @@ public interface SpellEffectTimersConfig extends Config
 			position = 2,
 			keyName = "showVengTimersOverlay",
 			name = "Show vengeance timers",
-			description = "Configures if vengeance timers are shown"
+			description = "Configures if vengeance cooldown timers are shown"
 	)
 	default boolean showVengTimersOverlay()
 	{
@@ -45,11 +69,11 @@ public interface SpellEffectTimersConfig extends Config
 
 	@ConfigItem(
 			position = 3,
-			keyName = "showVengActivatedOverlay",
-			name = "Show vengeance acivated",
-			description = "Configures if vengeance timers are shown"
+			keyName = "showVengActiveOverlay",
+			name = "Show vengeance active",
+			description = "Configures if vengeance active icon is shown when a player has the effects of vengeance"
 	)
-	default boolean showVengActivatedOverlay()
+	default boolean showVengActiveOverlay()
 	{
 		return true;
 	}
