+++ runelite-client/src/main/java/net/runelite/client/plugins/spelleffecttimers/SpellEffectTimersPlugin.java
@@ -1,3 +1,27 @@
+/*
+ * Copyright (c) 2019, Siraz, Jajack
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
 package net.runelite.client.plugins.spelleffecttimers;
 
 import com.google.inject.Provides;
@@ -9,6 +33,7 @@ import net.runelite.api.coords.WorldPoint;
 import net.runelite.api.events.*;
 import net.runelite.client.config.ConfigManager;
 import net.runelite.client.eventbus.Subscribe;
+import net.runelite.client.game.MapLocations;
 import net.runelite.client.plugins.Plugin;
 import net.runelite.client.plugins.PluginDescriptor;
 import net.runelite.client.plugins.PluginType;
@@ -68,6 +93,7 @@ public class SpellEffectTimersPlugin extends Plugin
 		// 1. Ahrim's full set effect spot anim overriding (id 400)
 		// 2. Elysian hit effect might also trigger this
 		// 3. Enchanted bolt effects
+		// Everything that overrides the freeze spot animation causes this issue
 
 		GameState gameState = client.getGameState();
 		if (gameState == GameState.LOGGING_IN || gameState == GameState.CONNECTION_LOST || gameState == GameState.HOPPING)
@@ -170,7 +196,7 @@ public class SpellEffectTimersPlugin extends Plugin
 	{
 		log.debug("Check teleblock triggered | animationId: {} | Tick: {}", actor.getSpotAnimation(), client.getGameCycle());
 
-		Player player = (Player)actor;
+		Player player = (Player) actor;
 		boolean magePray = false;
 		if (player.getOverheadIcon() != null)
 		{
